Move AuthUI usage from Authentication.kt to SplashActivity.kt
