+++ app/build.gradle
@@ -66,7 +66,11 @@ dependencies {
 
     // Kotlin
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android-ktx:1.3.9'
+
+    // Kotlin Coroutines
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.3.8"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.5"
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-play-services:1.3.5'
 
     // Android
     implementation 'androidx.core:core-ktx:1.3.2'
