+++ app/src/main/java/com/ojomono/ionce/firebase/Authentication.kt
@@ -1,16 +1,12 @@
 package com.ojomono.ionce.firebase
 
 // TODO: Avoid Android imports and move to separated module when needed for more UI platforms
-import android.content.Intent
 import android.net.Uri
 import android.util.Log
-import androidx.core.net.toUri
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Transformations
 import com.facebook.AccessToken
-import com.firebase.ui.auth.AuthUI
-import com.firebase.ui.auth.IdpResponse
 import com.google.android.gms.tasks.Task
 import com.google.firebase.auth.*
 import com.google.firebase.auth.ktx.actionCodeSettings
@@ -29,25 +25,15 @@ object Authentication {
     /***************/
 
     // Dynamic Links
-    private const val DL_URL_DOMAIN = "https://ionce.page.link"
-    private const val DL_FINISH_SIGN_UP = "/finishSignUp"
+    const val DL_URL_DOMAIN = "https://ionce.page.link"
+    const val DL_FINISH_SIGN_UP = "/finishSignUp"
     private const val DL_LINK_WITH_EMAIL = "/linkWithEmail"
-
-    // Photo Url Size
-    private const val PU_WANTED_SIZE = 400  // Not all values are supported by Twitter API!
-    private const val PU_GOOGLE_DEFAULT_SIZE = 96
-    private const val PU_GOOGLE_SIZE_COMPONENT = "s%d-c"
-    private const val PU_FACEBOOK_SIZE_COMPONENT = "?height=%d&access_token=%s"
-    private const val PU_TWITTER_SIZE_COMPONENT_DEFAULT = "_normal"
-    private const val PU_TWITTER_SIZE_COMPONENT_WANTED = "_%dx%d"
 
     /************/
     /** Fields **/
     /************/
 
-    // TODO move AuthUI to SplashActivity to avoid importing Intent here
-    // The Firebase Authentication and it's pre-built UI instances
-    private val authUI = AuthUI.getInstance()
+    // The Firebase Authentication
     private val auth = FirebaseAuth.getInstance().apply {
         // Update the LiveData every time the underlying token state changes.
         addAuthStateListener { _currentUser.value = currentUser }
@@ -84,107 +70,6 @@ object Authentication {
         Transformations.map(currentUser) {
             it?.providerData?.find { data -> data.providerId == PhoneAuthProvider.PROVIDER_ID }
         }
-
-    /*********************/
-    /** Sign in methods **/
-    /*********************/
-    // TODO this section will probably move to SplashActivity
-
-    /**
-     * Build an intent that will open the FirebaseUI sign-in screen. Enable email link directing to
-     * the APPLICATION_ID (main package). If the function is called after the email link
-     * was clicked, the link should be given as the [activityIntent] model.
-     */
-    fun buildSignInIntent(activityIntent: Intent? = null): Intent {
-
-        // Start building the sign-in Intent and email builders
-        val signInIntentBuilder = authUI.createSignInIntentBuilder()
-        val emailBuilder = AuthUI.IdpConfig.EmailBuilder()
-
-        // If the intent is given and AuthUI can handle it (email link clicked),
-        // state it in the sign-in Intent
-        if (activityIntent != null && AuthUI.canHandleIntent(activityIntent)) {
-            val link = activityIntent.data.toString()
-            signInIntentBuilder.setEmailLink(link)
-        }
-
-        // Enable email link sign in
-        val actionCodeSettings: ActionCodeSettings =
-            ActionCodeSettings.newBuilder()
-                .setAndroidPackageName(BuildConfig.APPLICATION_ID, true, null)
-                .setHandleCodeInApp(true) // This must be set to true
-                .setUrl(DL_URL_DOMAIN + DL_FINISH_SIGN_UP) // This URL needs to be whitelisted
-                .build()
-        emailBuilder.enableEmailLinkSignIn().setActionCodeSettings(actionCodeSettings)
-
-        // Choose authentication providers
-        val providers = arrayListOf(
-            AuthUI.IdpConfig.GoogleBuilder().build(),
-            AuthUI.IdpConfig.FacebookBuilder().build(),
-            AuthUI.IdpConfig.TwitterBuilder().build(),
-            emailBuilder.build(),
-            AuthUI.IdpConfig.PhoneBuilder().build()
-        )
-
-        // Return the built intent
-        return signInIntentBuilder.setAvailableProviders(providers)
-            .setLogo(R.drawable.app_logo) // Set logo drawable
-            .setTheme(R.style.AppTheme_SignInScreen) // Set theme
-            .build()
-    }
-
-    /**
-     * Handle the case of a successful sign-in.
-     */
-    fun handleSignInSucceeded(dataIntent: Intent?) {
-        val response = IdpResponse.fromResultIntent(dataIntent)
-
-        // For new created users, check if a photo is available from the auth provider
-        if (response?.isNewUser == true) {
-            // All users have a default FirebaseAuth provider data - we want to check which is the
-            // other one
-            currentUser.value?.providerData
-                ?.find { it.providerId != FirebaseAuthProvider.PROVIDER_ID }?.apply {
-                    val photoUrl = when (providerId) {
-                        GoogleAuthProvider.PROVIDER_ID ->
-                            // Replace "s96-c" to "s400-c" to get 400x400 image
-                            photoUrl.toString().replace(
-                                PU_GOOGLE_SIZE_COMPONENT.format(PU_GOOGLE_DEFAULT_SIZE),
-                                PU_GOOGLE_SIZE_COMPONENT.format(PU_WANTED_SIZE)
-                            )
-                        FacebookAuthProvider.PROVIDER_ID ->
-                            // TODO: Get facebook link (when Graph API will allow it again)
-                            // Add parameters: "?height=400&access_token=${response.idpToken}" to
-                            // get 400x400 image
-                            photoUrl.toString().plus(
-                                PU_FACEBOOK_SIZE_COMPONENT.format(PU_WANTED_SIZE, response.idpToken)
-                            )
-                        TwitterAuthProvider.PROVIDER_ID ->
-                            // TODO: Get Twitter screen_name (when getting Facebook link)
-                            // Replace "_normal" to "_400x400" to get 400x400 image
-                            photoUrl.toString().replace(
-                                PU_TWITTER_SIZE_COMPONENT_DEFAULT,
-                                PU_TWITTER_SIZE_COMPONENT_WANTED
-                                    .format(PU_WANTED_SIZE, PU_WANTED_SIZE)
-                            )
-                        else -> null
-                    }
-                    photoUrl?.let { updatePhotoUrl(it.toUri()) }
-                }
-        }
-    }
-
-    /**
-     * Handle the case of a failed sign-in. Response should be given in [dataIntent].
-     */
-    fun handleSignInFailed(dataIntent: Intent?) {
-        val response = IdpResponse.fromResultIntent(dataIntent)
-
-        // If response is null the user canceled the sign-in flow using the back button.
-        // Otherwise an error occurred:
-        // Use response.error.errorCode if a specific error handling is needed.
-        if (response != null) Log.e(TAG, response.error.toString())
-    }
 
     /***********************************/
     /** Invoke firebase tasks methods **/
